+++ doc/Movable.html
@@ -3,7 +3,7 @@
 <html lang="sv">
 <head>
 <!-- Generated by javadoc (15) on Sun Jan 24 17:34:30 CET 2021 -->
-<title>Model.Movable</title>
+<title>Model.Model.Movable</title>
 <meta http-equiv="Content-Type" content="text/html; charset=utf-8">
 <meta name="dc.created" content="2021-01-24">
 <meta name="description" content="declaration: interface: Model.Movable">
