+++ doc/Volvo240.html
@@ -3,7 +3,7 @@
 <html lang="sv">
 <head>
 <!-- Generated by javadoc (15) on Sun Jan 24 17:34:30 CET 2021 -->
-<title>Model.Volvo240</title>
+<title>Model.Model.Volvo240</title>
 <meta http-equiv="Content-Type" content="text/html; charset=utf-8">
 <meta name="dc.created" content="2021-01-24">
 <meta name="description" content="declaration: class: Model.Volvo240">
