+++ doc/index-files/index-11.html
@@ -105,7 +105,7 @@ loadScripts(document, 'script');</script>
 <dd>&nbsp;</dd>
 <dt><a href="../SpeedPane.html" title="class in &lt;Unnamed&gt;"><span class="type-name-link">SpeedPane</span></a> - Class in <a href="../package-summary.html">&lt;Unnamed&gt;</a></dt>
 <dd>&nbsp;</dd>
-<dt><span class="member-name-link"><a href="../SpeedPane.html#%3Cinit%3E(Model.Vehicle)">SpeedPane(Model.Vehicle)</a></span> - Constructor for class <a href="../SpeedPane.html" title="class in &lt;Unnamed&gt;">SpeedPane</a></dt>
+<dt><span class="member-name-link"><a href="../SpeedPane.html#%3Cinit%3E(Model.Model.Vehicle)">SpeedPane(Model.Vehicle)</a></span> - Constructor for class <a href="../SpeedPane.html" title="class in &lt;Unnamed&gt;">SpeedPane</a></dt>
 <dd>&nbsp;</dd>
 <dt><span class="member-name-link"><a href="../Presentation.html#start(javafx.stage.Stage)">start(Stage)</a></span> - Method in class <a href="../Presentation.html" title="class in &lt;Unnamed&gt;">Presentation</a></dt>
 <dd>&nbsp;</dd>
