+++ src/Model/Car.java
@@ -1,5 +1,7 @@
 package Model;
 
+import View.IPaintable;
+
 import java.awt.*;
 
 /**
@@ -7,7 +9,7 @@ import java.awt.*;
  */
 public abstract class Car extends Vehicle implements Transportable, IPaintable {
     /**
-     * Model.Car model
+     * Model.Model.Car model
      */
     private final String modelName;
 
@@ -35,7 +37,7 @@ public abstract class Car extends Vehicle implements Transportable, IPaintable {
         this.modelName = modelName;
         this.color = color;
         this.nrDoors = nrDoors;
-//        vehicle.setPosition(new Model.Vector2D(0, 0));
+//        vehicle.setPosition(new Model.Model.Vector2D(0, 0));
     }
     /**
      * Creates a representation of a car
