+++ src/Model/Container.java
@@ -8,7 +8,7 @@ import java.util.Deque;
  *
  * @param <T> The type stored in the container
  */
-public abstract class Container<T extends Transportable> /*implements Model.Transporter<T>*/ {
+public abstract class Container<T extends Transportable> /*implements Model.Model.Transporter<T>*/ {
     private final int capacity;
     private final double maxWidth;
     private final double maxLength;
@@ -41,7 +41,7 @@ public abstract class Container<T extends Transportable> /*implements Model.Tran
      */
     public void store(T item) {
         if (holder.size() >= capacity) {
-            throw new IllegalStateException("Model.Container full");
+            throw new IllegalStateException("Model.Model.Container full");
         }
         if (item.getLength() > maxLength) {
             throw new IllegalArgumentException("Item too long");
@@ -58,7 +58,7 @@ public abstract class Container<T extends Transportable> /*implements Model.Tran
 
     /**
      * Eject the item stored in holder.
-     * current item is decided by findItemToRemove(), which depends on if it's Model.LIFO or Model.FIFO.
+     * current item is decided by findItemToRemove(), which depends on if it's Model.Model.LIFO or Model.Model.FIFO.
      * @return gives a item stored in holder.
      */
     public T remove() {
@@ -84,13 +84,13 @@ public abstract class Container<T extends Transportable> /*implements Model.Tran
     }
 
     /**
-     * FindItemToRemove finds which object to remove, depending on Model.LIFO or Model.FIFO
+     * FindItemToRemove finds which object to remove, depending on Model.Model.LIFO or Model.Model.FIFO
      * @return type of container class.
      */
     protected abstract T findItemToRemove();
 
     /**
-     * findOffset to decide if the "direction" T should be Model.FIFO or Model.LIFO.
+     * findOffset to decide if the "direction" T should be Model.Model.FIFO or Model.Model.LIFO.
      * @return type of container class
      */
     protected abstract Vector2D findOffset();
