+++ src/Model/ITrim.java
@@ -1,7 +1,7 @@
 package Model;
 
 /**
- * Interface Model.ITrim is used to trim a car,
+ * Interface Model.Model.ITrim is used to trim a car,
  * findSpeedFactor should be multiplied with a trim factor.
  */
 public interface ITrim {
