+++ src/Model/ITurbo.java
@@ -1,7 +1,7 @@
 package Model;
 
 /**
- * Model.ITurbo is a interface to add a turbo to a Model.Car.
+ * Model.Model.ITurbo is a interface to add a turbo to a Model.Model.Car.
  */
 public interface ITurbo {
     void setTurboOn();
