+++ src/Model/Model.java
@@ -1,3 +1,5 @@
+package Model;
+
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 import java.util.ArrayList;
