+++ src/Model/ObserverHandler.java
@@ -1,5 +1,7 @@
 package Model;
 
+import View.IPaintable;
+
 import java.util.List;
 import java.util.ArrayList;
