+++ src/Model/Transportable.java
@@ -1,7 +1,7 @@
 package Model;
 
 /**
- * Model.Transportable interface enables an object to be transported, i.e. loaded into a container.
+ * Model.Model.Transportable interface enables an object to be transported, i.e. loaded into a container.
  */
 public interface Transportable {
     void setDirection(Vector2D direction);
