+++ src/Model/Transporter.java
@@ -1,7 +1,7 @@
 package Model;
 
 /**
- * Model.Transporter interface enables an object to act as a transporter.
+ * Model.Model.Transporter interface enables an object to act as a transporter.
  * @param <T> things to transport
  */
 public interface Transporter<T extends Transportable> {
