+++ src/Model/Vehicle.java
@@ -195,16 +195,16 @@ public abstract class Vehicle implements Movable {
     }
 
     /**
-     * Returns x and y coordinate of Model.Vehicle
+     * Returns x and y coordinate of Model.Model.Vehicle
      *
-     * @return Model.Vector2D(x, y)
+     * @return Model.Model.Vector2D(x, y)
      */
     public Vector2D getPosition() {
         return position;
     }
 
     /**
-     * Model.Vehicle width
+     * Model.Model.Vehicle width
      *
      * @return width of the vehicle
      */
@@ -213,7 +213,7 @@ public abstract class Vehicle implements Movable {
     }
 
     /**
-     * Model.Vehicle length
+     * Model.Model.Vehicle length
      *
      * @return length of the vehicle
      */
@@ -233,7 +233,7 @@ public abstract class Vehicle implements Movable {
     /**
      * Set current direction of the vehicle.
      *
-     * @param direction set a direction of length 1 in Model.Vector2D.
+     * @param direction set a direction of length 1 in Model.Model.Vector2D.
      */
     public void setDirection(Vector2D direction) {
         this.direction = direction;
