+++ src/Model/Volvo240.java
@@ -3,7 +3,7 @@ package Model;
 import java.awt.*;
 
 /**
- * A class representing a Model.Volvo240 car.
+ * A class representing a Model.Model.Volvo240 car.
  */
 public class Volvo240 extends Car implements Transportable, ITrim {
 
@@ -13,15 +13,15 @@ public class Volvo240 extends Car implements Transportable, ITrim {
     private final static double trimFactor = 1.25;
 
     /**
-     * Constructs a Model.Volvo240 object with (hardcoded) presets for modelName, color, enginePower and nrDoors.
+     * Constructs a Model.Model.Volvo240 object with (hardcoded) presets for modelName, color, enginePower and nrDoors.
      */
     public Volvo240(double width, double length) {
-        super(100, width, length, "Model.Volvo240", Color.black, 4);
+        super(100, width, length, "Model.Model.Volvo240", Color.black, 4);
     }
 
     /**
-     * Returns the speed factor of the Model.Volvo240.
-     * @return the speed factor of the Model.Volvo240
+     * Returns the speed factor of the Model.Model.Volvo240.
+     * @return the speed factor of the Model.Model.Volvo240
      */
     protected double findSpeedFactor() {
         return getEnginePower() * 0.01 * trimFactor;
