+++ src/View/DrawPanel.java
@@ -9,6 +9,9 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.ArrayList;
 import java.util.Map;
+import Model.Vector2D;
+
+
 
 // This panel represent the animated part of the view with the car images.
 
@@ -27,9 +30,9 @@ public class DrawPanel extends JPanel{
     }
 
     private void bindCars() {
-        bind("Model.Saab95", "pics/Saab95.jpg");
-        bind("Model.Volvo240", "pics/Volvo240.jpg");
-        bind("Model.Scania", "pics/Scania.jpg");
+        bind("Model.Model.Saab95", "pics/Saab95.jpg");
+        bind("Model.Model.Volvo240", "pics/Volvo240.jpg");
+        bind("Model.Model.Scania", "pics/Scania.jpg");
     }
 
     public void bind(String name, String imageFileName) {
@@ -52,12 +55,12 @@ public class DrawPanel extends JPanel{
     }
 
     /*
-    public void paintCars(List<? extends Model.IPaintable> paintables){
+    public void paintCars(List<? extends View.IPaintable> paintables){
         Graphics graphics = getGraphics();
         super.paintComponent(graphics);
 
-        for (Model.IPaintable paintable : paintables) {
-            Model.Vector2D pos = paintable.getPosition();
+        for (View.IPaintable paintable : paintables) {
+            Model.Model.Vector2D pos = paintable.getPosition();
             BufferedImage image = images.get(paintable.getModelName());
             graphics.drawImage(image, (int)pos.getX(), (int)pos.getY(), null);
         }
