+++ test/TestVehicle.java
@@ -2,6 +2,7 @@
 
 import org.junit.Before;
 import org.junit.Test;
+import Model.*;
 
 
 import static org.junit.Assert.*;
