+++ test/TestWorkshop.java
@@ -3,6 +3,7 @@ import org.junit.Test;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
+import Model.*;
 
 public class TestWorkshop {
     Workshop<Volvo240> bilia;
